feat: redirect profile route to profile-settings
